requirements.txt
torch>=1.12.0
torchvision>=0.13.0
transformers>=4.21.0
timm>=0.6.7
scikit-learn>=1.1.0
matplotlib>=3.5.0
seaborn>=0.11.0
numpy>=1.21.0
Pillow>=8.3.0
tqdm>=4.64.0
lime>=0.2.0
shap>=0.41.0